Fixet mit lort

+++ P3_Projekt_WPF/P3_Projekt_WPF/Classes/Product/TempProduct.cs
@@ -64,7 +64,7 @@ namespace P3_Projekt_WPF.Classes
         public override void CreateFromRow(Row Table)
         {
             ID = Convert.ToInt32(Table.Values[0]);
-            SalePrice = Convert.ToInt32(Table.Values[1]);
+            SalePrice = Convert.ToDecimal(Table.Values[1]);
             Description = Table.Values[2];
             _resolved = Convert.ToBoolean(Table.Values[3]);
         }
